- Made Apply Leave form into a shared component
