+++ src/app/manager/add-leave/add-leave.component.scss
@@ -66,127 +66,3 @@ $border-radius-sm: 6px;
   text-align: center;
 }
 
-// Form
-.leave-form {
-  .form-row {
-    display: flex;
-    gap: 1rem;
-
-    .form-group {
-      flex: 1;
-    }
-  }
-
-  .form-group {
-    margin-bottom: 1.5rem;
-
-    label {
-      display: block;
-      font-size: 0.875rem;
-      font-weight: 500;
-      color: $text-primary;
-      margin-bottom: 0.5rem;
-    }
-
-    input,
-    textarea {
-      width: 100%;
-      padding: 0.75rem 1rem;
-      font-size: 1rem;
-      line-height: 1.5;
-      color: $text-primary;
-      background-color: $background-white;
-      border: 2px solid $border-color;
-      border-radius: $border-radius-sm;
-      transition: all 0.2s ease-in-out;
-      box-sizing: border-box;
-
-      &:focus {
-        outline: none;
-        border-color: $border-focus;
-        box-shadow: 0 0 0 3px rgba(79, 124, 255, 0.1);
-      }
-
-      &.ng-invalid.ng-touched {
-        border-color: $error-color;
-        box-shadow: 0 0 0 3px rgba(229, 62, 62, 0.1);
-      }
-    }
-
-    textarea {
-      resize: vertical;
-      min-height: 100px;
-    }
-  }
-
-  // Submit button
-  .form-actions {
-    margin-top: 2rem;
-
-    .btn-primary {
-      //width: 100%;
-      padding: 0.875rem 1.5rem;
-      font-size: 1rem;
-      font-weight: 500;
-      color: white;
-      background-color: $primary-color;
-      border: none;
-      border-radius: $border-radius-sm;
-      cursor: pointer;
-      transition: all 0.2s ease-in-out;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 0.5rem;
-
-      //&:hover:not(:disabled) {
-      //  background-color: $primary-hover;
-      //  transform: translateY(-1px);
-      //  box-shadow: $shadow-md;
-      //}
-      //
-      //&:active:not(:disabled) {
-      //  transform: translateY(0);
-      //}
-      //
-      //&:disabled {
-      //  background-color: $primary-disabled;
-      //  cursor: not-allowed;
-      //  transform: none;
-      //  box-shadow: none;
-      //}
-
-      .loading-spinner {
-        width: 1rem;
-        height: 1rem;
-        border: 2px solid rgba(255, 255, 255, 0.3);
-        border-top: 2px solid white;
-        border-radius: 50%;
-        animation: spin 1s linear infinite;
-      }
-    }
-  }
-}
-
-// Loading spinner animation
-@keyframes spin {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
-}
-
-// Responsive design
-@media (max-width: 640px) {
-  :host {
-    padding: 1rem;
-  }
-
-  .form-container {
-    padding: 1.5rem;
-  }
-
-  .form-header {
-    .form-title {
-      font-size: 1.5rem;
-    }
-  }
-}
