+++ src/app/manager/add-leave/add-leave.component.ts
@@ -6,18 +6,18 @@ import { LeaveService } from '../../shared-data/leaveapplication.service';
 import { CurrentUserService } from '../../shared-data/currentUserService';
 import { PaginatedLeaveApplication } from '../../shared-data/paginated-leave-application';
 import {UsersService} from '../../shared-data/users.service';
+import {AddLeaveFormComponent} from '../../shared-components/add-leave-form/add-leave-form.component';
 
 @Component({
   selector: 'app-manager-add-leave',
   standalone: true,
-  imports: [ReactiveFormsModule, CommonModule],
+  imports: [ReactiveFormsModule, CommonModule, AddLeaveFormComponent],
   templateUrl: './add-leave.component.html',
   styleUrls: ['./add-leave.component.scss']
 })
-export class ManagerAddLeaveComponent implements OnInit{
-  leaveForm: FormGroup;
-  isSubmitting = false;
+export class ManagerAddLeaveComponent{
   availableLeave = 0;
+  isSubmitting = false;
 
   constructor(
     private readonly router: Router,
@@ -25,98 +25,35 @@ export class ManagerAddLeaveComponent implements OnInit{
     private readonly currentUserService: CurrentUserService,
     private readonly usersService: UsersService
   ) {
-    const user = this.currentUserService.getCurrentUser();
-    this.availableLeave = user?.remainingCredits ?? 0;
-
-    this.leaveForm = new FormGroup({
-      startDate: new FormControl('', Validators.required),
-      endDate: new FormControl('', Validators.required),
-      totalDays: new FormControl(0, Validators.required),
-      remarks: new FormControl('', [Validators.required, Validators.minLength(5)])
-    });
-
     effect(() => {
       const user = this.currentUserService.getCurrentUser();
       if (user) {
-        this.availableLeave = user.remainingCredits;
+        this.availableLeave = user.remainingCredits ?? 0;
       }
     });
   }
 
-  ngOnInit() {
-    this.leaveForm.get('startDate')?.valueChanges.subscribe(() => this.calculateTotalDays());
-    this.leaveForm.get('endDate')?.valueChanges.subscribe(() => this.calculateTotalDays());
-  }
+  saveLeaveApplication(formValue: any) {
+    const user = this.currentUserService.getCurrentUser();
+    if (!user) return;
 
-  calculateTotalDays() {
-    const start = this.leaveForm.get('startDate')?.value;
-    const end = this.leaveForm.get('endDate')?.value;
+    this.isSubmitting = true;
+    this.leaveService.applyLeave(user.id, formValue).subscribe({
+      next: (response: PaginatedLeaveApplication) => {
+        alert('Leave request submitted successfully!');
 
-    if (!start || !end) {
-      this.leaveForm.get('totalDays')?.setValue(0);
-      return;
-    }
-
-    const startDate = new Date(start);
-    const endDate = new Date(end);
-
-    if (endDate < startDate) {
-      this.leaveForm.get('totalDays')?.setValue(0);
-      return;
-    }
-
-    let days = 0;
-    let current = new Date(startDate);
-
-    while (current <= endDate) {
-      const dayOfWeek = current.getDay();
-      if (dayOfWeek !== 0 && dayOfWeek !== 6) {
-        days++;
-      }
-      current.setDate(current.getDate() + 1);
-    }
-
-    this.leaveForm.get('totalDays')?.setValue(days);
-  }
-  saveLeaveApplication() {
-    if (this.leaveForm.valid) {
-      const totalDays = this.leaveForm.get('totalDays')?.value;
-      if (totalDays > this.availableLeave) {
-        alert(`You only have ${this.availableLeave} leave days remaining.`);
-        return;
-      }
-
-      this.isSubmitting = true;
-      const requestBody = this.leaveForm.getRawValue();
-      const user = this.currentUserService.getCurrentUser();
-
-      if (user) {
-        this.leaveService.applyLeave(user.id, requestBody).subscribe({
-          next: (response: PaginatedLeaveApplication) => {
-            console.log('Leave request submitted:', response);
-            alert('Leave request submitted successfully!');
-
-            // update available leave immediately in UI
-            this.availableLeave = Math.max(this.availableLeave - totalDays, 0);
-
-            this.usersService.getUserById(user.id).subscribe(updatedUser => {
-              this.currentUserService.setCurrentUser(updatedUser); // update global state
-              this.availableLeave = updatedUser.remainingCredits ?? 0;
-            });
-
-            // reset form
-            this.leaveForm.reset({ totalDays: 0 });
-            this.router.navigate(['manager/add-leave']);
-          },
-          error: (err) => {
-            console.error('Failed to submit leave request:', err);
-            alert('Failed to submit leave request.');
-          },
-          complete: () => {
-            this.isSubmitting = false;
-          }
+        this.usersService.getUserById(user.id).subscribe(updatedUser => {
+          this.currentUserService.setCurrentUser(updatedUser);
+          this.availableLeave = updatedUser.remainingCredits ?? 0;
         });
+      },
+      error: (err) => {
+        console.error('Failed to submit leave request:', err);
+        alert('Failed to submit leave request.');
+      },
+      complete: () => {
+        this.isSubmitting = false;
       }
-    }
+    });
   }
 }
