+++ src/app/shared-components/add-leave-form/add-leave-form.component.html
@@ -1,0 +1,54 @@
+<form [formGroup]="leaveForm" (ngSubmit)="onSubmit()" class="leave-form">
+  <div class="form-row">
+    <div class="form-group">
+      <label for="startDate" class="form-label">Start Date</label>
+      <input
+        id="startDate"
+        type="date"
+        formControlName="startDate"
+        class="form-input"
+      />
+    </div>
+
+    <div class="form-group">
+      <label for="endDate" class="form-label">End Date</label>
+      <input
+        id="endDate"
+        type="date"
+        formControlName="endDate"
+        class="form-input"
+      />
+    </div>
+  </div>
+
+  <div class="form-group">
+    <label for="totalDays" class="form-label">Total Days</label>
+    <input
+      id="totalDays"
+      type="number"
+      formControlName="totalDays"
+      class="form-input"
+      readonly
+    />
+  </div>
+
+  <div class="form-group">
+    <label for="remarks" class="form-label">Reason</label>
+    <textarea
+      id="remarks"
+      formControlName="remarks"
+      class="form-input"
+      placeholder="Please provide reason for leave"
+    ></textarea>
+  </div>
+
+  <div class="form-actions">
+    <button
+      type="submit"
+      class="btn-primary"
+      [disabled]="leaveForm.invalid || isSubmitting"
+    >
+      {{ isSubmitting ? 'Submitting Request...' : 'Submit Request' }}
+    </button>
+  </div>
+</form>
